requirements.txt
# Enhanced requirements.txt with LLM integration
numpy>=1.21.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
networkx>=2.8.0
typing-extensions>=4.0.0
flask>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0

# Environment variable management
python-dotenv>=1.0.0

# LLM Integration Dependencies
openai>=1.0.0
asyncio-throttle>=1.0.2  # For rate limiting API calls

# Optional: For better async handling
aiohttp>=3.8.0
aiofiles>=0.8.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=4.0.0